Update sip-X.md
Changes to the structure of SIPs to provide more clarity for authors as to the intended purpose of each section and to break out the specification for further clarity. Specification has been expanded and subsections added to improve the flow of information for the reader and help guide the author on how to communicate the intent behind the SIP more clearly.